fix: add handler for getStoriesByIds to fix errors
