+++ lib/storyblok/StoryblokContext.js
@@ -1,5 +1,6 @@
 import React, { createContext, useContext, useState } from 'react';
 import getStoriesByContentType from '@/lib/storyblok/helpers/getStoriesByContentType';  
+import getStoriesByIds from './helpers/getStoriesByIds';
 
 const StoryblokContext = createContext();
 
@@ -11,11 +12,16 @@ export const StoryblokProvider = ({ children, stories, globalSettings }) => {
     return getStoriesByContentType(contextStories, contentType);
   };
 
+  const getStoriesByIdsHandler = (uuids) => {
+    return getStoriesByIds(contextStories, uuids);
+  };
+
   return (
     <StoryblokContext.Provider value={{ 
       stories: contextStories, 
-      globalSettings: contextGlobalSettings, 
-      getStoriesByContentType: getStoriesByContentTypeHandler
+      globalSettings: contextGlobalSettings,
+      getStoriesByContentType: getStoriesByContentTypeHandler,
+      getStoriesByIds: getStoriesByIdsHandler,
     }}>
       {children}
     </StoryblokContext.Provider>
